--- src/main/java/kr/hhplus/be/server/reservation/interfaces/web/dto/response/ReservationTokenIssueResponse.java
@@ -1,5 +1,0 @@
-package kr.hhplus.be.server.reservation.interfaces.web.dto.response;
-
-public record ReservationTokenIssueResponse(String token) {
-
-}
