--- src/main/java/kr/hhplus/be/server/reservation/interfaces/web/dto/response/ReservationTokenStatusResponse.java
@@ -1,5 +1,0 @@
-package kr.hhplus.be.server.reservation.interfaces.web.dto.response;
-
-public record ReservationTokenStatusResponse(String status, Long rank) {
-
-}
