+++ src/main/java/kr/hhplus/be/server/reservation/interfaces/web/dto/response/reservation/ReservationTokenIssueResponse.java
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.reservation.interfaces.web.dto.response.reservation;
+
+public record ReservationTokenIssueResponse(String token) {
+
+}
