+++ src/main/java/kr/hhplus/be/server/reservation/interfaces/web/dto/response/reservation/ReservationTokenStatusResponse.java
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.reservation.interfaces.web.dto.response.reservation;
+
+public record ReservationTokenStatusResponse(String status, Long rank) {
+
+}
